weaviate-connection/requirements.txt
weaviate-client==4.4.4
requests==2.31.0
